#149: Initial refactoring: moved platform independent things to core-shared
