+++ core-linux/bin/app/neutralino.png
