--- core-linux/bin/app/settings-browser.json
@@ -1,5 +1,0 @@
-{
-    "appname" : "myapp",
-    "appport" : "8080",
-    "mode" : "browser"
-}
