--- core-linux/bin/app/settings-cloud.json
@@ -1,8 +1,0 @@
-{
-    "appname" : "myapp",
-    "appport" : "8080",
-    "mode" : "cloud",
-    "cloud" : {
-        "blacklist" : ["os.runCommand"]
-    }
-}
