+++ core-linux/build.sh
@@ -26,7 +26,7 @@ if [ -e bin/neutralino ]; then
     rm bin/neutralino
 fi
 
-g++ -I ../core-shared -std=c++17 ../core-shared/log.cpp src/resources.cpp src/Buffer.cpp src/Handler.cpp src/requestparser.cpp src/Socket.cpp src/functions.cpp src/main.cpp src/router.cpp src/core/filesystem/filesystem.cpp src/settings.cpp src/core/os/os.cpp src/core/computer/computer.cpp src/core/debug/debug.cpp src/auth/authbasic.cpp src/ping/ping.cpp src/core/storage/storage.cpp src/core/app/app.cpp src/cloud/privileges.cpp -pthread -std=c++17 -DWEBVIEW_GTK=1 `pkg-config --cflags --libs gtk+-3.0 webkit2gtk-4.0` -o bin/neutralino -no-pie -lstdc++fs
+g++ -I ../core-shared -std=c++17 ../core-shared/log.cpp ../core-shared/resources.cpp src/server/Buffer.cpp src/server/Handler.cpp src/server/requestparser.cpp src/server/Socket.cpp ../core-shared/helpers.cpp src/main.cpp src/server/router.cpp src/core/filesystem/filesystem.cpp src/platform/linux.cpp ../core-shared/settings.cpp src/core/os/os.cpp src/core/computer/computer.cpp src/core/debug/debug.cpp ../core-shared/auth/authbasic.cpp ../core-shared/ping/ping.cpp src/core/storage/storage.cpp src/core/app/app.cpp ../core-shared/cloud/privileges.cpp -pthread -std=c++17 -DWEBVIEW_GTK=1 `pkg-config --cflags --libs gtk+-3.0 webkit2gtk-4.0` -o bin/neutralino -no-pie -lstdc++fs
 
 if [ -e bin/neutralino ]; then
     echo "Neutralino binary is compiled in to bin/neutralino"
