+++ core-shared/auth/authbasic.h
