+++ core-shared/cloud/privileges.h
@@ -20,21 +20,10 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#include "../lib/json/json.hpp"
 
 using namespace std;
-using json = nlohmann::json;
-extern bool loadResFromDir;
 
-namespace settings {
-    json getSettings();
-    json getOptions();
-    void setOption(string key, string value);
-    string getFileContent(string filename);
-    string getFileContentBinary(string filename);
-    string getGlobalVars();
-    string getCurrentDir();
-    void setGlobalArgs(json args);
-    string joinAppPath(string filename);
+namespace privileges {
+    vector<string> getBlacklist();
+    bool checkPermission(string func);
 }
-
