+++ neutralinojs
@@ -1,0 +1,35 @@
+[editor]
+line_wrapping=false
+line_break_column=72
+auto_continue_multiline=true
+
+[file_prefs]
+final_new_line=true
+ensure_convert_new_lines=false
+strip_trailing_spaces=false
+replace_tabs=false
+
+[indentation]
+indent_width=4
+indent_type=0
+indent_hard_tab_width=8
+detect_indent=false
+detect_indent_width=false
+indent_mode=2
+
+[project]
+name=neutralinojs
+base_path=/home/shalithasuranga/Projects/neutralinojs
+description=
+file_patterns=
+
+[long line marker]
+long_line_behaviour=1
+long_line_column=72
+
+[files]
+current_page=0
+FILE_NAME_0=1669;C++;0;EUTF-8;0;1;0;%2Fhome%2Fshalithasuranga%2FProjects%2Fneutralinojs%2Fcore-linux%2Fsrc%2Fresources.cpp;0;4
+
+[VTE]
+last_dir=/home/shalithasuranga
